+++ cpplox/include/lox/debug.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "chunk.h"
+#include "lox/chunk.h"
 
 void disassembleChunk(Chunk* chunk, std::string_view name);
-int disassembleInstruction(Chunk* chunk, int offset);
+size_t disassembleInstruction(Chunk* chunk, size_t offset);
